+++ app/src/main/res/navigation/nav_graph.xml
@@ -27,7 +27,7 @@
         tools:layout="@layout/settings_fragment" />
     <fragment
         android:id="@+id/searchFragment"
-        android:name="github.karchx.wiki.SearchFragment"
+        android:name="github.karchx.wiki.ui.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/search_fragment" />
     <fragment
